Data Access

The gridded fields used in the ECCO examples can be retrieved from ecco-group.org and, for the ECCOv4r2 estimate, from Harvard Dataverse or zenodo.org. Two monthly climatologies (ECCOv4r2 and OCCA) are also readily available using the Julia artifact system as explained below. These can be relatively large files, compared to the package codes, so they are handled lazily (only downloaded when needed).

Artifact path	File Type	Download Method

ECCOclim_path	NetCDF	lazy, by variable, dataverse

OCCAclim_path	NetCDF	lazy, by variable, dataverse

MITPROFclim_path	binary	lazy, whole, zenodo

ECCOdiags_path	JLD2	lazy, whole, zenodo

Basic Usage

For ECCO :

using OceanStateEstimation, MeshArrays
γ=GridSpec("LatLonCap",MeshArrays.GRID_LLC90)
tmp=OceanStateEstimation.get_ecco_files(γ,"ETAN")

For OCCA :

using OceanStateEstimation
get_occa_variable_if_needed("SIarea")
readdir(OCCAclim_path)

Functions Reference

OceanStateEstimation.ECCO.standard_analysis_setup — Method

ECCO.standard_analysis_setup(pth0)

Create a run folder where folder pth0 will be linked.

Folder pth0 (of type String) should be the path to the user's ECCO data folder.

using Revise
using OceanStateEstimation, Pkg

#pth0="nctiles_monthly" #edit path as needed
pth=ECCO.standard_analysis_setup(pth0)

Pkg.activate(pth)
Pkg.instantiate()

#include("ECCO_standard_loop.jl")

OceanStateEstimation.ECCO_helpers.parameters — Method

parameters(pth0::String,sol0::String,list0,index)

Prepare parameter NamedTuple for use in ECCO_diagnostics.driver.

OceanStateEstimation.ECCO_io.read_monthly — Method

read_monthly(P,nam,t)

Read record t for variable nam from file locations specified via P.

Depending on nam this will call read_monthly_default, read_monthly_SSH, read_monthly_MHT, or read_monthly_BSF.

OceanStateEstimation.ECCO_diagnostics.driver — Method

driver(P)

Call main loop from computation, files, etc specified via P.

Depending on P this will call main_clim, main_glo, main_zonmean, main_overturn, main_MHT, or main_trsp.