Data Access

The gridded fields used in the ECCO examples can be retrieved from ecco-group.org and, for the ECCOv4r2 estimate, from Harvard Dataverse or zenodo.org. Two monthly climatologies (ECCOv4r2 and OCCA) are also readily available using the Julia artifact system as explained below. These can be relatively large files, compared to the package codes, so they are handled lazily (only downloaded when needed).

Artifact path	File Type	Download Method

ECCOclim_path	NetCDF	lazy, by variable, dataverse

OCCAclim_path	NetCDF	lazy, by variable, dataverse

MITPROFclim_path	binary	lazy, whole, zenodo

ECCOdiags_path	JLD2	lazy, whole, zenodo

Basic Usage

For ECCO :

using OceanStateEstimation, MeshArrays
γ=GridSpec("LatLonCap",MeshArrays.GRID_LLC90)
tmp=OceanStateEstimation.get_ecco_files(γ,"ETAN")

For OCCA :

using OceanStateEstimation
get_occa_variable_if_needed("SIarea")
readdir(OCCAclim_path)

Functions Reference

OceanStateEstimation.ECCO.standard_analysis_setup — Method

ECCO.standard_analysis_setup(pth0::String)

Create temporary run folder pth where data folder pth0 will be linked.

Data folder pth0 should be the path to ECCO data.

For example:

using OceanStateEstimation, Pkg
pth_in=ECCOclim_path
pth=ECCO.standard_analysis_setup(pth_in)

The Project.toml file found in pth provides an environment ready for ECCO analyses.

This environment can be activated and instantiated:

Pkg.activate(pth)
Pkg.instantiate()

OceanStateEstimation.ECCO_helpers.parameters — Method

parameters(P0,params)

Prepare parameter NamedTuple for use in ECCO_diagnostics.driver.

P1=parameters(P0,p)

is faster than e.g. parameters(pth,"r2",p) as grid, etc get copied from P0 to P1.

OceanStateEstimation.ECCO_helpers.parameters — Method

parameters(pth0::String,sol0::String,params)

Prepare parameter NamedTuple for use in ECCO_diagnostics.driver.

For example, to compute zonal mean temperatures at level 5:

p=(calc = "zonmean", nam = "THETA", lev = 5)
pth=ECCO.standard_analysis_setup(ECCOclim_path)
P0=parameters(pth,"r2",p)

or, from a predefined list:

list0=ECCO_helpers.standard_list_toml("")
pth=ECCO.standard_analysis_setup(ECCOclim_path)
P1=parameters(pth,"r2",list0[1])

OceanStateEstimation.ECCO_diagnostics.driver — Method

driver(P)

Call main computation loop as specified by parameters P.

The main computation loop choice depends on the P parameter values. Methods include:

main_clim

main_glo

main_zonmean

main_overturn

main_MHT

main_trsp

OceanStateEstimation.ECCO_io.read_monthly — Method

read_monthly(P,nam,t)

Read record t for variable nam from file locations specified via parameters P.

The method used to read nam is selected based on nam's value. Methods include:

read_monthly_default

read_monthly_SSH

read_monthly_MHT

read_monthly_BSF