Data Access

The gridded fields used in the ECCO examples can be retrieved from ecco-group.org and, for the ECCOv4r2 estimate, from Harvard Dataverse or zenodo.org. Two monthly climatologies (ECCOv4r2 and OCCA) are also readily available using the Julia artifact system as explained below. These can be relatively large files, compared to the package codes, so they are handled lazily (only downloaded when needed).

Artifact path	File Type	Download Method

ECCOclim_path	NetCDF	lazy, by variable, dataverse

OCCAclim_path	NetCDF	lazy, by variable, dataverse

MITPROFclim_path	binary	lazy, whole, zenodo

ECCOdiags_path	JLD2	lazy, whole, zenodo

Basic Usage

For ECCO :

using OceanStateEstimation, MeshArrays
γ=GridSpec("LatLonCap",MeshArrays.GRID_LLC90)
tmp=OceanStateEstimation.get_ecco_files(γ,"ETAN")

For OCCA :

using OceanStateEstimation
get_occa_variable_if_needed("SIarea")
readdir(OCCAclim_path)

Functions Reference

OceanStateEstimation.CBIOMESclim_download — Method

CBIOMESclim_download()

Download lazy artifact to CBIOMESclim_path.

OceanStateEstimation.ECCOdiags_add — Method

ECCOdiags_add(nam::String)

Add data to the ECCOdiagspath folder. Known options for nam include "release1", "release3", "release4", and "interpcoeffs". Note that "release2" is the estimate that's readily donwloaded by ECCOdiags_download().

OceanStateEstimation.ECCOdiags_download — Method

ECCOdiags_download()

Download lazy artifact to ECCOdiags_path.

OceanStateEstimation.MITPROFclim_download — Method

MITPROFclim_download()

Download lazy artifact to MITPROFclim_path.

OceanStateEstimation.dataverse_lists — Method

dataverse_lists(lst::String)

Read and derive lists (ID,name,URL) from csv file (ID,name) and return as tuple

lists=dataverse_lists(lst)

OceanStateEstimation.get_ecco_files — Function

get_ecco_files(γ::gcmgrid,v::String,t=1)

using MeshArrays, OceanStateEstimation
γ=GridSpec("LatLonCap",MeshArrays.GRID_LLC90)
tmp=get_ecco_files(γ,"oceQnet")

OceanStateEstimation.get_ecco_variable_if_needed — Method

get_ecco_variable_if_needed(v::String)

Download ECCO output for variable v to ECCOclim_path if needed

OceanStateEstimation.get_ecco_velocity_if_needed — Method

get_ecco_velocity_if_needed()

Download ECCO output for u,v,w to ECCOclim_path if needed

OceanStateEstimation.get_from_dataverse — Method

get_from_dataverse(lst::String,nam::String,pth::String)

using OceanStateEstimation, CSV, DataFrames
pth=dirname(pathof(OceanStateEstimation))
lst=joinpath(pth,"../examples/OCCA_climatology.csv")
nams=CSV.read(lst,DataFrame)
[get_from_dataverse(lst,nam,OCCAclim_path) for nam in nams.name[:]]

OceanStateEstimation.get_occa_variable_if_needed — Method

get_occa_variable_if_needed(v::String)

Download OCCA output for variable v to OCCAclim_path if needed

OceanStateEstimation.get_occa_velocity_if_needed — Method

get_occa_velocity_if_needed()

Download OCCA output for u,v,w to OCCAclim_path if needed