Intro

Climatologies are readily accessed using the Scratch.jl artifact system via OceanStateEstimation.jl as explained below.

Use Examples

ECCO

ECCO climatology files can downloaded using get_ecco_files. These files are for version 4 release 2, on the native model grid.

using OceanStateEstimation, MeshArrays
γ=GridSpec("LatLonCap",MeshArrays.GRID_LLC90)
tmp=OceanStateEstimation.get_ecco_files(γ,"ETAN")

5×1 MeshArrays.gcmarray{Float64, 2, Matrix{Float64}}:
 [NaN NaN … NaN NaN; NaN NaN … NaN NaN; … ; NaN NaN … NaN NaN; NaN NaN … NaN NaN]
 [NaN NaN … NaN NaN; NaN NaN … NaN NaN; … ; NaN NaN … NaN NaN; NaN NaN … NaN NaN]
 [NaN NaN … NaN NaN; NaN NaN … NaN NaN; … ; NaN NaN … NaN NaN; NaN NaN … NaN NaN]
 [NaN NaN … NaN NaN; NaN NaN … NaN NaN; … ; NaN NaN … NaN NaN; NaN NaN … NaN NaN]
 [NaN NaN … NaN NaN; NaN NaN … NaN NaN; … ; NaN NaN … NaN NaN; NaN NaN … NaN NaN]

Precomputed quantities shown in ECCO_standard_plots.jl can be downloaded separately.

OceanStateEstimation.ECCOdiags_add("release2")
readdir(ScratchSpaces.ECCO)

2-element Vector{String}:
 "ECCOv4r2_analysis"
 "ETAN"

OCCA

get_occa_variable_if_needed("SIarea")
readdir(ScratchSpaces.OCCA)

1-element Vector{String}:
 "SIarea.0406clim.nc"

CBIOMES

To retrieve the CBIOMES climatology, in the julia REPL for example :

OceanStateEstimation.CBIOMESclim_download()
readdir(ScratchSpaces.CBIOMES)

1-element Vector{String}:
 "CBIOMES-global-alpha-climatology.nc"

And the files, now found in ScratchSpaces.CBIOMES, can then be read using other libraries.

using NCDatasets
fil_out=joinpath(ScratchSpaces.CBIOMES,"CBIOMES-global-alpha-climatology.nc")
nc=NCDataset(fil_out,"r")
keys(nc)

21-element Vector{String}:
 "SST"
 "land"
 "t"
 "lat"
 "lon"
 "climatology_bounds"
 "tim"
 "Chl"
 "EuphoticDepth"
 "MLD"
 ⋮
 "Rrs412"
 "Rrs443"
 "Rrs490"
 "Rrs510"
 "Rrs555"
 "Rrs670"
 "SSS"
 "TKE"
 "WindSpeed"

Path Names

Gridded fields are mostly retrieved from Harvard Dataverse. These can be relatively large files, compared to the package codes, so they are handled lazily (only downloaded when needed). Precomputed diagnostics have also been archived on zenodo.org.

Artifact Name	File Type	Download Method

ScratchSpaces.ECCO	NetCDF	lazy, by variable, dataverse

ScratchSpaces.ECCO	JLD2	lazy, whole, zenodo

ScratchSpaces.MITprof	binary	lazy, whole, zenodo

ScratchSpaces.OCCA	NetCDF	lazy, by variable, dataverse

ScratchSpaces.CBIOMES	NetCDF	lazy, whole, zenodo

Functions Reference

OceanStateEstimation.downloads.CBIOMESclim_download — Method

CBIOMESclim_download()

Download lazy artifact to scratch space.

OceanStateEstimation.downloads.ECCOdiags_add — Method

ECCOdiags_add(nam::String)

Add data to the scratch space folder. Known options for nam include "release1", "release2", "release3", "release4", and "interp_coeffs".

OceanStateEstimation.downloads.MITPROFclim_download — Method

MITPROFclim_download()

Download lazy artifact to scratch space.

OceanStateEstimation.downloads.get_ecco_files — Function

get_ecco_files(γ::gcmgrid,v::String,t=1)

using MeshArrays, OceanStateEstimation, MITgcmTools
γ=GridSpec("LatLonCap",MeshArrays.GRID_LLC90)
tmp=OceanStateEstimation.get_ecco_files(γ,"oceQnet")

OceanStateEstimation.downloads.get_ecco_variable_if_needed — Method

get_ecco_variable_if_needed(v::String)

Download ECCO output for variable v to scratch space if needed

OceanStateEstimation.downloads.get_ecco_velocity_if_needed — Method

get_ecco_velocity_if_needed()

Download ECCO output for u,v,w to scratch space if needed

OceanStateEstimation.downloads.get_occa_variable_if_needed — Method

get_occa_variable_if_needed(v::String)

Download OCCA output for variable v to scratch space if needed

OceanStateEstimation.downloads.get_occa_velocity_if_needed — Method

get_occa_velocity_if_needed()

Download OCCA output for u,v,w to scratch space if needed