Path Variables

The gridded fields used in the ECCO examples can be retrieved from ecco-group.org and, for the ECCOv4r2 estimate, from Harvard Dataverse or zenodo.org. Two monthly climatologies (ECCOv4r2 and OCCA) are also readily available using the Scratch.jl artifact system as explained below. These can be relatively large files, compared to the package codes, so they are handled lazily (only downloaded when needed).

Artifact Name	File Type	Download Method

ECCO clim	NetCDF	lazy, by variable, dataverse

ECCO diags	JLD2	lazy, whole, zenodo

MITprof clim	binary	lazy, whole, zenodo

OCCA clim	NetCDF	lazy, by variable, dataverse

CBIOMES clim	NetCDF	lazy, whole, zenodo

Basic Usage

ECCO

ECCO climatology files can downloaded using get_ecco_files. These files are for version 4 release 2, on the native model grid.

using OceanStateEstimation, MeshArrays
γ=GridSpec("LatLonCap",MeshArrays.GRID_LLC90)
tmp=OceanStateEstimation.get_ecco_files(γ,"ETAN")

Precomputed quantities shown in ECCO_standard_plots.jl can be downloaded separately.

OceanStateEstimation.ECCOdiags_add("release2")
OceanStateEstimation.ECCOdiags_add("interp_coeffs")

OCCA

using OceanStateEstimation
get_occa_variable_if_needed("SIarea")
readdir(ScratchSpaces.OCCA)

CBIOMES

To retrieve the CBIOMES climatology, in the julia REPL for example :

using OceanStateEstimation
OceanStateEstimation.CBIOMESclim_download()

And the files, now found in ScratchSpaces.CBIOMES, can then be read using other libraries.

using NCDatasets
fil_out=joinpath(ScratchSpaces.CBIOMES,"CBIOMES-global-alpha-climatology.nc")
nc=NCDataset(fil_out,"r")

Functions Reference

OceanStateEstimation.downloads.CBIOMESclim_download — Method

CBIOMESclim_download()

Download lazy artifact to scratch space.

OceanStateEstimation.downloads.ECCOdiags_add — Method

ECCOdiags_add(nam::String)

Add data to the scratch space folder. Known options for nam include "release1", "release2", "release3", "release4", and "interp_coeffs".

OceanStateEstimation.downloads.MITPROFclim_download — Method

MITPROFclim_download()

Download lazy artifact to scratch space.

OceanStateEstimation.downloads.dataverse_lists — Method

dataverse_lists(lst::String)

Read and derive lists (ID,name,URL) from csv file (ID,name) and return as tuple

lists=dataverse_lists(lst)

OceanStateEstimation.downloads.get_ecco_files — Function

get_ecco_files(γ::gcmgrid,v::String,t=1)

using MeshArrays, OceanStateEstimation
γ=GridSpec("LatLonCap",MeshArrays.GRID_LLC90)
tmp=get_ecco_files(γ,"oceQnet")

OceanStateEstimation.downloads.get_ecco_variable_if_needed — Method

get_ecco_variable_if_needed(v::String)

Download ECCO output for variable v to scratch space if needed

OceanStateEstimation.downloads.get_ecco_velocity_if_needed — Method

get_ecco_velocity_if_needed()

Download ECCO output for u,v,w to scratch space if needed

OceanStateEstimation.downloads.get_from_dataverse — Method

get_from_dataverse(lst::String,nam::String,pth::String)

using OceanStateEstimation, CSV, DataFrames
pth=dirname(pathof(OceanStateEstimation))
lst=joinpath(pth,"../examples/OCCA_climatology.csv")
nams=CSV.read(lst,DataFrame)
[get_from_dataverse(lst,nam,ScratchSpaces.OCCA) for nam in nams.name[:]]

OceanStateEstimation.downloads.get_occa_variable_if_needed — Method

get_occa_variable_if_needed(v::String)

Download OCCA output for variable v to scratch space if needed

OceanStateEstimation.downloads.get_occa_velocity_if_needed — Method

get_occa_velocity_if_needed()

Download OCCA output for u,v,w to scratch space if needed