The workflow presented here is as follows.

set up for running analyses of ECCO estimates.

run one computation loop on the ECCO monthly files.

Climatology.ECCO.standard_analysis_setup — Function

ECCO.standard_analysis_setup(pth0::String)

Create temporary run folder pth where data folder pth0 will be linked.

Data folder pth0 should be the path to ECCO data.

For example:

using Climatology, Pkg
pth=ECCO.standard_analysis_setup(ScratchSpaces.ECCO)

The Project.toml file found in pth provides an environment ready for ECCO analyses.

This environment can be activated and instantiated:

Pkg.activate(pth)
Pkg.instantiate()

Here is an example of parameters P to compute zonal mean temperatures at level 5.

Climatology.ECCO_helpers.parameters — Function

parameters(P0,params)

Prepare parameter NamedTuple for use in ECCO_diagnostics.driver.

P1=parameters(P0,p)

is faster than e.g. parameters(pth,"r2",p) as grid, etc get copied from P0 to P1.

parameters(pth0::String,sol0::String,params)

Prepare parameter NamedTuple for use in ECCO_diagnostics.driver.

For example, to compute zonal mean temperatures at level 5:

p=(calc = "zonmean", nam = "THETA", lev = 5)
pth=ECCO.standard_analysis_setup(ScratchSpaces.ECCO)
P0=ECCO_helpers.parameters(pth,"r2",p)

or, from a predefined list:

list0=ECCO_helpers.standard_list_toml("")
pth=ECCO.standard_analysis_setup(ScratchSpaces.ECCO)
P1=ECCO_helpers.parameters(pth,"r2",list0[1])

The computation loop, over all months, can then be carried out as follows.

Climatology.ECCO_diagnostics.driver — Function

driver(P)

Call main computation loop as specified by parameters P.

The main computation loop choice depends on the P parameter values. Methods include:

main_clim

main_glo

main_zonmean

main_overturn

main_MHT

main_trsp

Climatology.ECCO_io.read_monthly — Method

read_monthly(P,nam,t)

Read record t for variable nam from file locations specified via parameters P.

The method used to read nam is selected based on nam's value. Methods include:

read_monthly_default

read_monthly_SSH

read_monthly_MHT

read_monthly_BSF